[Major] Reorganisation of files, notably slicing ModLoader.cs into several files
