+++ ModUtils/AsmUtils.cs
@@ -1,11 +1,117 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using Serilog;
 using UndertaleModLib;
+using UndertaleModLib.Decompiler;
 using UndertaleModLib.Models;
 
 namespace ModShardLauncher
 {
+    public static class AsmUtils
+    {
+        public static string GetAssemblyString(string fileName)
+        {
+            try 
+            {
+                UndertaleCode originalCode = CodeUtils.GetUMTCodeFromFile(fileName);
+                return originalCode.Disassemble(ModLoader.Data.Variables, ModLoader.Data.CodeLocals.For(originalCode));
+            }
+            catch(Exception ex) 
+            {
+                Log.Error(ex, "Something went wrong");
+                throw;
+            }
+        }
+        public static void SetAssemblyString(string codeAsString, string fileName)
+        {
+            try 
+            {
+                UndertaleCode originalCode = CodeUtils.GetUMTCodeFromFile(fileName);
+                originalCode.Replace(Assembler.Assemble(codeAsString, ModLoader.Data));
+            }
+            catch(Exception ex) 
+            {
+                Log.Error(ex, "Something went wrong");
+                throw;
+            }
+        }
+        public static void InsertAssemblyString(string codeAsString, string fileName, int position)
+        {
+            try 
+            {
+                Log.Information(string.Format("Trying insert assembly in: {0}", fileName.ToString()));
+
+                List<string>? originalCode = GetAssemblyString(fileName).Split("\n").ToList();
+                originalCode.Insert(position, codeAsString);
+                SetAssemblyString(string.Join("\n", originalCode), fileName);
+
+                Log.Information(string.Format("Patched function with InsertDisassemblyCode: {0}", fileName.ToString()));
+            }
+            catch(Exception ex) 
+            {
+                Log.Error(ex, "Something went wrong");
+                throw;
+            }
+        }
+        public static void ReplaceAssemblyString(string codeAsString, string fileName, int position)
+        {
+            try 
+            {
+                Log.Information(string.Format("Trying replace assembly in: {0}", fileName.ToString()));
+
+                List<string>? originalCode = GetAssemblyString(fileName).Split("\n").ToList();
+                originalCode[position] = codeAsString;
+                SetAssemblyString(string.Join("\n", originalCode), fileName);
+
+                Log.Information(string.Format("Patched function with ReplaceDisassemblyCode: {0}", fileName.ToString()));
+            }
+            catch(Exception ex) 
+            {
+                Log.Error(ex, "Something went wrong");
+                throw;
+            }
+        }
+        public static void ReplaceAssemblyString(string codeAsString, string fileName, int start, int len)
+        {
+            try 
+            {
+                Log.Information(string.Format("Trying replace assembly in: {0}", fileName.ToString()));
+
+                List<string>? originalCode = GetAssemblyString(fileName).Split("\n").ToList();
+                originalCode[start] = codeAsString;
+                for (int i = 1; i < Math.Min(len, originalCode.Count - start); i++) {
+                    originalCode[start + i] = "";
+                }
+
+                SetAssemblyString(string.Join("\n", originalCode), fileName);
+
+                Log.Information(string.Format("Patched function with ReplaceDisassemblyCode: {0}", fileName.ToString()));
+            }
+            catch(Exception ex) 
+            {
+                Log.Error(ex, "Something went wrong");
+                throw;
+            }
+        }
+        public static void InjectAssemblyInstruction(string name, Func<IEnumerable<UndertaleInstruction>, IEnumerable<UndertaleInstruction>> patch)
+        {
+            try 
+            {
+                Log.Information(string.Format("Trying patch assembly in: {0}", name.ToString()));
+
+                UndertaleCode originalCode = CodeUtils.GetUMTCodeFromFile(name);
+                originalCode.Replace(patch(originalCode.Instructions).ToList());
+
+                Log.Information(string.Format("Patched function with PatchDisassemblyCode: {0}", name.ToString()));
+            }
+            catch(Exception ex) 
+            {
+                Log.Error(ex, "Something went wrong");
+                throw;
+            }
+        }
+    }
     public static class AssemblyWrapper
     {
         public static UndertaleInstruction.Reference<UndertaleVariable> CreateRefVariable(string name, UndertaleInstruction.InstanceType instanceType)
