+++ ModUtils/RoomUtils.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+using UndertaleModLib.Models;
+
+namespace ModShardLauncher
+{
+    public static class RoomUtils
+    {
+        public static IEnumerable<UndertaleRoom> GetRooms()
+        {
+            return ModLoader.Data.Rooms;
+        }
+    }
+}
